fixed MSVC compile warnings

+++ tutorials/user_geometry/user_geometry_device.cpp
@@ -347,7 +347,7 @@ void sphereIntersectFuncN(const int* valid,
       RTCRayN_v(rays,N,ui) = 0.0f;
       RTCRayN_tfar(rays,N,ui) = t0;
       RTCRayN_geomID(rays,N,ui) = sphere.geomID;
-      RTCRayN_primID(rays,N,ui) = item;
+      RTCRayN_primID(rays,N,ui) = (unsigned int)item;
       const Vec3fa Ng = ray_org+t0*ray_dir-sphere.p;
       RTCRayN_Ng_x(rays,N,ui) = Ng.x;
       RTCRayN_Ng_y(rays,N,ui) = Ng.y;
@@ -359,7 +359,7 @@ void sphereIntersectFuncN(const int* valid,
       RTCRayN_v(rays,N,ui) = 0.0f;
       RTCRayN_tfar(rays,N,ui) = t1;
       RTCRayN_geomID(rays,N,ui) = sphere.geomID;
-      RTCRayN_primID(rays,N,ui) = item;
+      RTCRayN_primID(rays,N,ui) = (unsigned int)item;
       const Vec3fa Ng = ray_org+t1*ray_dir-sphere.p;
       RTCRayN_Ng_x(rays,N,ui) = Ng.x;
       RTCRayN_Ng_y(rays,N,ui) = Ng.y;
@@ -744,7 +744,7 @@ void renderTileStandardStream(int taskIndex,
     /* initialize ray */
     RTCRay& primary = primary_stream[N];
     primary.org = Vec3fa(camera.xfm.p);
-    primary.dir = Vec3fa(normalize(x*camera.xfm.l.vx + y*camera.xfm.l.vy + camera.xfm.l.vz));
+    primary.dir = Vec3fa(normalize((float)x*camera.xfm.l.vx + (float)y*camera.xfm.l.vy + camera.xfm.l.vz));
     mask = 1; { // invalidates inactive rays
       primary.tnear = mask ? 0.0f         : (float)(pos_inf);
       primary.tfar  = mask ? (float)(inf) : (float)(neg_inf);
